+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add Chat Popout functionality for attack and item chat cards
 - Attack chat cards now show the attack's icon
 - Attack and item chat cards now show the combat round they were shown, if in combat
+- Add containers to list of targets for the 'Give Item' action on actor sheets
 
 ## 0.75.7